requirements.txt
httpx
fastapi
pydantic
aiogram
python-dotenv
alembic
asyncpg
asyncio
psycopg2-binary
uvicorn
sqlalchemy
